Move changes stream into PoiLogController
This became necessary for a couple of reasons:

- The changes stream in its very first version only operated on Pois so
  it was obvious to put it in the NodesController
- Now all data is gathered from a separate model namely PoiLog
- The plan is to implement pagination using the established mechanisms used
  already in other controllers
- The NodesController operates on Pois but we use PoiLogs
